requirements.txt
matplotlib==3.7.1
numpy==1.24.2
Pillow==9.4.0
PyQt5==5.15.9
scipy==1.10.1
torch==1.13.1 --extra-index-url https://download.pytorch.org/whl/cu116
torch-mtcnn==0.0.7
torchvision==0.14.1
tqdm==4.65.0
